Fixed potential casting issue reported by CI

+++ packages/dear-imgui/cpp/app/include/widget/plot_candlestick.h
@@ -1,3 +1,5 @@
+#include <cstdint>
+
 #include "shared.h"
 #include "styled_widget.h"
 
@@ -60,7 +62,7 @@ public:
             return snprintf(buff, size,"0");
         }
 
-        return snprintf(buff, size, "%.*f", reinterpret_cast<int>(decimalPlaces), value);
+        return snprintf(buff, size, "%.*f", reinterpret_cast<intptr_t>(decimalPlaces), value);
     };
 
     bool HasCustomWidth() override;
